requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
django-taggit==1.4.0
django-widget-tweaks==1.4.8
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
